# task03/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Blog Post Backend
  description: |-
    This is a sample blog post backend based on the OpenAPI 3.0 specification. 

    Some useful links:
    - [Zalando Api Guidelines](https://opensource.zalando.com/restful-api-guidelines)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:8080
...
paths:
...
  /blog-posts:
    get:
      tags:
        - Blog Posts
      summary: Get all blogposts
      description: Returns all blogposts
      operationId: getBlogPosts
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPosts'
            application/xml:
              schema:
                $ref: '#/components/schemas/BlogPosts'
    post:
      tags:
        - Blog Posts
      summary: Add a new blog post
      description: Add a new blog post
      operationId: addBlockPost
      requestBody:
        description: Create a new blog post
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlogPostRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/BlogPostRequest'
        required: true
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlogPost'
            application/xml:
              schema:
                $ref: '#/components/schemas/BlogPost'
...
components:
  schemas:
    BlogPostRequest:
      type: object
      properties:
        locked:
          type: boolean
          example: true
        status:
          type: string
          description: Blog post status
          example: PUBLISHED
          enum:
            - DRAFT
            - PUBLISHED
        # https://opensource.zalando.com/restful-api-guidelines/#118
        blog_post_author_id:
          type: integer
          format: int64
          example: 7
        content:
          type: string
      xml:
        name: BlogPostRequest
    BlogPost:
      allOf:
        - $ref: '#/components/schemas/BlogPostRequest'
        - type: object
          properties:
            blog_post_id:
              type: integer
              format: int64
              example: 10
      xml:
        name: BlogPost
    BlogPosts:
      type: object
      properties:
        # https://opensource.zalando.com/restful-api-guidelines/#120
        blog_posts:
          type: array
          items:
            $ref: '#/components/schemas/BlogPost'
      xml:
        name: blog-posts
...